Remove alternative from files names
